requirements.txt
flask>=2.2
flask-socketio>=5.3
python-socketio>=5.9
python-engineio>=4.5
nidaqmx>=0.8
numpy>=1.24
paho-mqtt>=1.6
nptdms>=1.10
pyserial>=3.5
pymodbus>=3.6
